+++ src/content/otrs-profile.js
@@ -1,0 +1,4 @@
+export function getOtrsUserId(document) {
+	const idInput = document.getElementById("CurrentUserID")
+	return idInput?.value
+}